Fix one error and all warnings on MSVC 2026
The error is "C1001: Internal compiler error." in uci.cpp on line 370 of
master.  For some reason the compiler can't handle the
std::size(hashfullAges) inside the lambda.  Older MSVC versions had no
problem.

Most of the warnings are due to implicit type conversions "conversion
from type A to type B, possible loss of data"
many of which have been present for a while.

No functional change

+++ src/timeman.cpp
@@ -92,7 +92,7 @@ void TimeManagement::init(Search::LimitsType& limits,
 
     // If less than one second, gradually reduce mtg
     if (scaledTime < 1000)
-        centiMTG = scaledTime * 6.000;
+        centiMTG = int(scaledTime * 6.000);
 
     // Make sure timeLeft is > 0 since we may use it as a divisor
     TimePoint timeLeft =
